% Generated by roxygen2 (4.0.2): do not edit by hand
\name{data.qc}
\alias{data.qc}
\title{!! NEEDS WORK !!}
\usage{
data.qc(data = NULL, out.dir = NULL, Cond.cutoff = NULL,
  Ci.cutoff = NULL, Tleaf.cutoff = NULL)
}
\arguments{
\item{data}{input licor data}

\item{out.dir}{output director for QC data and info}

\item{Cond.cutoff}{cutoff for low conductance. set previously}

\item{Ci.cutoff}{cutoff for nonsensical Cis}

\item{Tleaf.cutoff}{cutoff for individual Tleaf variation from mean}
}
\description{
QA/QC functions
}
\author{
Shawn P. Serbin
}

